apps/backend/pyproject.toml
[tool.coverage.run]
branch = true
source = [ "ocr" ]

[tool.coverage.report]
exclude_lines = ['if TYPE_CHECKING:']
show_missing = true

[tool.pytest.ini_options]
addopts = "--cov --cov-report html:'../../coverage/apps/backend/html' --cov-report xml:'../../coverage/apps/backend/coverage.xml' --html='../../reports/apps/backend/unittests/html/index.html' --junitxml='../../reports/apps/backend/unittests/junit.xml'"

[tool.poetry]
name = "ocr"
version = "1.0.0"
description = "Automatically generated by Nx."
authors = [ ]
license = 'Proprietary'
readme = 'README.md'

  [[tool.poetry.packages]]
  include = "ocr"

  [tool.poetry.dependencies]
  python = ">=3.9,<3.11"
  pytesseract = "^0.3.13"
  pillow = "^10.4.0"
  tensorflow = "^2.17.0"
  opencv-python = "^4.10.0.84"

  [tool.poetry.group.dev.dependencies]
  autopep8 = "2.0.2"
  flake8 = "6.0.0"
  pytest = "7.3.1"
  pytest-sugar = "0.9.7"
  pytest-cov = "4.1.0"
  pytest-html = "3.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
